backend/requirements.txt
# Workspace version: 1.46.32
ipykernel==6.29.5
databutton==0.38.34
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.9
openai
beautifulsoup4
requests
asyncpg
Pillow
slowapi
redis